requirements.txt
pymupdf==1.23.5
pdfminer.six==20221105
scikit-learn==1.3.2
spacy==3.7.2
sentence-transformers==2.2.2
huggingface-hub==0.14.1

